+++ braindecode/datasets/__init__.py
@@ -9,6 +9,7 @@ from .base import (
 )
 from .bcicomp import BCICompetitionIVDataset4
 from .bids import BIDSDataset, BIDSEpochsDataset
+from .collate import pad_channels_collate
 from .chb_mit import CHBMIT
 from .mne import create_from_mne_epochs, create_from_mne_raw
 from .moabb import BNCI2014_001, HGD, MOABBDataset
@@ -42,4 +43,5 @@ __all__ = [
     "SleepPhysionetChallenge2018",
     "create_from_X_y",
     "BCICompetitionIVDataset4",
+    "pad_channels_collate",
 ]